Small fixes #5

+++ api/src/main/groovy/io/builders/demo/dtcc/domain/lsmbatch/event/LsmExecutedDltEvent.groovy
@@ -4,7 +4,7 @@ import io.builders.demo.core.event.DltEvent
 
 class LsmExecutedDltEvent extends DltEvent {
 
-    Integer batchId
+    BigInteger batchId
     List<LsmTransaction> transactions
 
 }
